fix(httpclient): add http-module and interceptors
