+++ src/app/shared/http-module/http.module.ts
@@ -1,0 +1,19 @@
+import { HTTP_INTERCEPTORS, HttpClientModule } from '@angular/common/http';
+import { NgModule } from '@angular/core';
+import { AuthInterceptor } from './auth.interceptors';
+import { HttpClientService } from './http.service';
+
+@NgModule({
+  imports: [
+    HttpClientModule
+  ],
+  providers: [
+    HttpClientService,
+    {
+      provide: HTTP_INTERCEPTORS,
+      useClass: AuthInterceptor,
+      multi: true
+    }
+  ]
+})
+export class PlannerHttpClientModule {}
